services_thumb_1768160172.477000/requirements.txt
functions-framework==3.*  
google-cloud-storage
pdf2image
Pillow
PyMuPDF
